+++ tests/testthat/t2/R/t2.R
@@ -33,6 +33,13 @@ S7::method(another_s7_generic, an_s7_class) <- function(x) "foo"
 another_s3_generic <- S7::new_external_generic("t1", "another_s3_generic", "x")
 S7::method(another_s3_generic, an_s7_class) <- function(x) "foo"
 
+# A local generic with a method dispatched on a class from a soft dep (t1).
+# Must register lazily, after t1 is loaded.
+#' @export
+own_generic <- S7::new_generic("own_generic", "x")
+S7::method(own_generic, S7::new_external_class("t1", "Another S7 Class")) <-
+  function(x) "from t1 class"
+
 
 .onLoad <- function(libname, pkgname) {
   S7::methods_register()